PicoDst_Local_Test didn't work.
Put it together locally and realized all the includes were local ("")
and needed to be changed to standard (<>). Problem for later. Tried to
run from root directly on the test.picoDst.root file and that failed.
Tried to run similarly from RCF and that also failed curiously. Dead in
the water in terms of running the code myself for now.