Issue with corrrupted root files. Not resolved but hopefully mitigated.
When running threads in parallel, the first set of simulations would end
up with corrupted root files (histograms or canvases). It didn't matter
what they were called or what was in them. Writing to file at same time
is my best guess at the cause. To hopefully mitigate the problem, I
added a sleep command in loop queueing simulations. This will hopefully
stagger the write times and make simultaneous writing unlikely. Not sure
this is the issue but seems to behave now. Not a great solution.